refactor: move github_issue_comments.py into integrations/github/
Per the codebase conventions, integration-specific modules should live
under their vendor package. Move github_issue_comments.py into
integrations/github/ and update all imports accordingly.

No compatibility shim is left — per AGENTS.md, forwarding-only modules
must be removed in the same change.

Closes #3575
